remove and/or refactor FirebaseInstanceIdService references - deprecated
complete todo 1 & 2

add insertSquawk method

add displayNotification method.  Currently, it just has the notification, which does not open the app when touched.

As a side note, it only shows the duck icon on the JellyPro notifications.  All 4 units receive the Squawks, but the N4 isn't receiving the notifications now.  It was for a while, but I don't recall what the change may have been when it quit.